Refactor custom-server kode
