+++ server/tsconfig.json
@@ -1,15 +1,14 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "strict": false,
+    "strict": true,
     "forceConsistentCasingInFileNames": true,
     "esModuleInterop": true,
     "module": "CommonJS",
     "moduleResolution": "node",
-    "resolveJsonModule": true,
     "isolatedModules": true,
     "incremental": true,
-    "downlevelIteration": true,
+    "baseUrl": "./",
     "lib": [
       "dom",
       "esnext"
